Homework 3 edited
