+++ src/components/reviewForm/component.jsx
@@ -1,5 +1,6 @@
 import { useReducer } from "react";
 import { Counter } from "../counter/component";
+import { useCount } from "../../hooks/useCount";
 
 const FORM_INIT = {
     username: "",
@@ -26,6 +27,8 @@ function reducer(state, {type, payload}) {
 export const ReviewForm = () => {
     const [form, dispatch] = useReviewForm(FORM_INIT);
     const {username, review} = form;
+    
+    const { count, decrement, increment } = useCount({minValue: 1, maxValue: 5});
 
     return (
         <div>
@@ -36,7 +39,7 @@ export const ReviewForm = () => {
             <input type="text" value={review} onChange={(event) => {dispatch({ type: "setReview", payload: event.target.value })}} />
             <br/>
             <label>Rating</label>
-            <Counter min={1} max={5}/>
+            <Counter count={count} increment={increment} decrement={decrement}/>
             <br/>
             <button onClick={() => {dispatch({ type: "clear" })}}>Save</button>
         </div>
